API_ApplicationSondage/api/requirements.txt
uvicorn
pydantic
bcrypt
fastapi
mysql-connector-python
python-dotenv